tests/submission/Infraware/Forms/contents/Forms/fieldset_elements.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>FieldSet_elements</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
		<fieldset id="input_field">
			<p><input type=hidden name="custname"></p>
			<p><input type=hidden name="custtel"></p>
			<p><input type=hidden name="custemail"></p>
 		</fieldset>
	</form>
	<script>

		var field = document.getElementById("input_field");

		try
		{
			var ele = field.elements;

			test(function() {
				assert_equals(ele.length, 3, "elements attribute is not correct.");
			});	
		}
		catch (e)
		{
			test(function() {
				assert_unreached("Error is raised.");
			});
		}

	</script>

 </body>
</html>
